Add <leader>w for more control over :write calls
<leader>w calls a function which instead of just writing current buffer,
it first prompts for file path+name first

The initial text gets placed into the command-line at the bottom to
allow user to edit it there. Pressing <CR> will cause vim to write the
file, change buffer to be that window then delete the old file. This
behaviour is *not* equivalent to :w %:p<TAB>.